+++ src/main/java/org/example/springhomework003/service/ServiceImplement/venueServiceImpl.java
@@ -23,7 +23,7 @@ public class venueServiceImpl implements VenueService {
     }
 
     @Override
-    public Venue getVenueById(Integer id) {
+    public Venue getVenueById(Long id) {
         return venueRepository.getVenueById(id);
     }
